requirements.txt
asgiref==3.3.4
Django==3.2.1
mtcnn-opencv==1.0.2
numpy==1.21.0
opencv-python==4.5.1.48
paho-mqtt==1.5.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
